Setup Dagger dependencies and add base top-level modules
To set up Dagger on this project I had to create some base classes to
make our life easy.

So we'll have a `BaseActivity` to our futures activities inherit it.
And is the same for our presentres they will have to inherit the
`BasePresenter` to not break the chain.

Also, I'm adding two scopes to be used in our application to make sure
we'll have single instances as long it exists. They are
`ActivityScope` and `FragmentScope`.

This adding too the `ActivityScope` and `FragmentScope` to be used to
make sure we'll have single instances as long its scopes exists.