Setup base methods and the `SchedulerProvider` to work with Rx
On this commit, I'm adding two functions to apply schedules to our Rx
sources easy and they are `applySingleSchedulers()` and
`applyObservableSchedulers()`.

How does it work? Both receive `SchedulerProvider` as a parameter and
this `schedulerProvider` implementation is responsible for know who
the `workerThread()` and `postWorkerThread()` are. This will
transformer feeding it an Observable/Single of one type and it'll
return an Observable/Single of another. That's exactly the same as
calling a series of operators inline.

How to use these transformers? This is used with `.compose()`.
```Kotlin
Single.just("test")
  .compose(applySingleSchedulers(schedulerProvider))
  .subscribe()
  ...
```

That's it!